refactor: Move BindingItem to separate file
