+++ src/Uno.UI/DataBinding/BindingPath.cs
@@ -22,7 +22,7 @@ using System.Diagnostics.CodeAnalysis;
 namespace Uno.UI.DataBinding
 {
 	[DebuggerDisplay("Path={_path} DataContext={_dataContext}")]
-	internal class BindingPath : IDisposable, IValueChangedListener
+	internal partial class BindingPath : IDisposable, IValueChangedListener
 	{
 		private static List<IPropertyChangedRegistrationHandler> _propertyChangedHandlers = new List<IPropertyChangedRegistrationHandler>(2);
 		private readonly string _path;
@@ -521,515 +521,6 @@ namespace Uno.UI.DataBinding
 		}
 		#endregion
 
-		private sealed class BindingItem : IBindingItem, IDisposable
-		{
-			private delegate void PropertyChangedHandler(object? previousValue, object? newValue, bool shouldRaiseValueChanged);
-
-			private ManagedWeakReference? _dataContextWeakStorage;
-
-			private readonly SerialDisposable _propertyChanged = new SerialDisposable();
-			private bool _disposed;
-			private readonly DependencyPropertyValuePrecedences? _precedence;
-			private readonly object? _fallbackValue;
-			private readonly bool _allowPrivateMembers;
-			private ValueGetterHandler? _valueGetter;
-			private ValueGetterHandler? _precedenceSpecificGetter;
-			private ValueGetterHandler? _substituteValueGetter;
-			private ValueSetterHandler? _valueSetter;
-			private ValueSetterHandler? _localValueSetter;
-			private ValueSetterHandler? _animationFillingValueSetter;
-			private ValueUnsetterHandler? _valueUnsetter;
-			private ValueUnsetterHandler? _animationFillingValueUnsetter;
-
-			private Type? _dataContextType;
-
-			public BindingItem(BindingItem next, string property, object fallbackValue) :
-				this(next, property, fallbackValue, null, false)
-			{
-			}
-
-			internal BindingItem(BindingItem? next, string property, object? fallbackValue, DependencyPropertyValuePrecedences? precedence, bool allowPrivateMembers)
-			{
-				Next = next;
-				PropertyName = property;
-				_precedence = precedence;
-				_fallbackValue = fallbackValue;
-				_allowPrivateMembers = allowPrivateMembers;
-			}
-
-			public object? DataContext
-			{
-				get => _dataContextWeakStorage?.Target;
-				set
-				{
-					if (!_disposed)
-					{
-						// Historically, Uno was processing property changes using INPC. Since the inclusion of DependencyObject
-						// values changes are now filtered by DependencyProperty updates, making equality updates at this location
-						// detrimental to the use of INPC events processing.
-						// In case of an INPC, the bindings engine must reevaluate the path completely from the raising point, regardless
-						// of the reference being changed.
-						if (FeatureConfiguration.Binding.IgnoreINPCSameReferences && !DependencyObjectStore.AreDifferent(DataContext, value))
-						{
-							return;
-						}
-
-						var weakDataContext = WeakReferencePool.RentWeakReference(this, value);
-						SetWeakDataContext(weakDataContext);
-					}
-				}
-			}
-
-			internal void SetWeakDataContext(ManagedWeakReference? weakDataContext, bool transferReferenceOwnership = false)
-			{
-				var previousStorage = _dataContextWeakStorage;
-
-				_dataContextWeakStorage = weakDataContext;
-				OnDataContextChanged();
-
-				// Return the reference to the pool after it's been released from the next BindingItem instances.
-				// Failing to do so makes the reference change without the bindings knowing about it,
-				// making the reference comparison always equal.
-				Uno.UI.DataBinding.WeakReferencePool.ReturnWeakReference(this, previousStorage);
-			}
-
-			public BindingItem? Next { get; }
-			public string PropertyName { get; }
-
-			public IValueChangedListener? ValueChangedListener { get; set; }
-
-			public object? Value
-			{
-				get
-				{
-					return GetSourceValue();
-				}
-				set
-				{
-					SetValue(value);
-				}
-			}
-
-			/// <summary>
-			/// Sets the value using the <see cref="_precedence"/>
-			/// </summary>
-			/// <param name="value">The value to set</param>
-			private void SetValue(object? value)
-			{
-				BuildValueSetter();
-				SetSourceValue(_valueSetter!, value);
-			}
-
-			/// <summary>
-			/// Sets the value using the <see cref="DependencyPropertyValuePrecedences.Local"/>
-			/// </summary>
-			/// <param name="value">The value to set</param>
-			public void SetLocalValue(object value)
-			{
-				BuildLocalValueSetter();
-				SetSourceValue(_localValueSetter!, value);
-			}
-
-			public void SetAnimationFillingValue(object value)
-			{
-				BuildAnimationFillingValueSetter();
-				SetSourceValue(_animationFillingValueSetter!, value);
-			}
-
-			public Type? PropertyType
-			{
-				get
-				{
-					if (DataContext != null)
-					{
-						return BindingPropertyHelper.GetPropertyType(_dataContextType!, PropertyName, _allowPrivateMembers);
-					}
-					else
-					{
-						return null;
-					}
-				}
-			}
-
-			internal object? GetPrecedenceSpecificValue()
-			{
-				BuildPrecedenceSpecificValueGetter();
-
-				return GetSourceValue(_precedenceSpecificGetter!);
-			}
-
-			internal object? GetSubstituteValue()
-			{
-				BuildSubstituteValueGetter();
-
-				return GetSourceValue(_substituteValueGetter!);
-			}
-
-			private bool _isDataContextChanging;
-
-			private void OnDataContextChanged()
-			{
-				if (DataContext != null)
-				{
-					ClearCachedGetters();
-					if (_propertyChanged.Disposable != null)
-					{
-#if !HAS_EXPENSIVE_TRYFINALLY // Try/finally incurs a very large performance hit in mono-wasm - https://github.com/dotnet/runtime/issues/50783
-						try
-#endif
-						{
-							_isDataContextChanging = true;
-							_propertyChanged.Disposable = null;
-						}
-#if !HAS_EXPENSIVE_TRYFINALLY // Try/finally incurs a very large performance hit in mono-wasm - https://github.com/dotnet/runtime/issues/50783
-						finally
-#endif
-						{
-							_isDataContextChanging = false;
-						}
-					}
-
-					_propertyChanged.Disposable = SubscribeToPropertyChanged();
-
-					RaiseValueChanged(Value);
-
-					if (Next != null)
-					{
-						Next.DataContext = Value;
-					}
-				}
-				else
-				{
-					if (Next != null)
-					{
-						Next.DataContext = null;
-					}
-					RaiseValueChanged(null);
-
-					_propertyChanged.Disposable = null;
-				}
-			}
-
-			private void OnPropertyChanged(object? previousValue, object? newValue, bool shouldRaiseValueChanged)
-			{
-				if (_isDataContextChanging && newValue is UnsetValue)
-				{
-					// We're in a "resubscribe" scenario when the DataContext is provided a new non-null value, so we don't need to
-					// pass through the DependencyProperty.UnsetValue.
-					// We simply discard this update.
-					return;
-				}
-
-				if (Next != null)
-				{
-					Next.DataContext = newValue;
-				}
-
-				if (shouldRaiseValueChanged && previousValue != newValue)
-				{
-					RaiseValueChanged(newValue);
-				}
-			}
-
-			private void ClearCachedGetters()
-			{
-				var currentType = DataContext!.GetType();
-
-				if (_dataContextType != currentType && _dataContextType != null)
-				{
-					_valueGetter = null;
-					_precedenceSpecificGetter = null;
-					_substituteValueGetter = null;
-					_localValueSetter = null;
-					_valueSetter = null;
-					_valueUnsetter = null;
-				}
-
-				_dataContextType = currentType;
-			}
-
-			private void BuildValueSetter()
-			{
-				if (_valueSetter == null && _dataContextType != null)
-				{
-					_valueSetter = BindingPropertyHelper.GetValueSetter(
-						_dataContextType,
-						PropertyName,
-						convert: true,
-						precedence: _precedence ?? DependencyPropertyValuePrecedences.Local
-					);
-					if (_precedence == null)
-					{
-						_localValueSetter = _valueSetter;
-					}
-				}
-			}
-
-			private void BuildLocalValueSetter()
-			{
-				if (_localValueSetter == null && _dataContextType != null)
-				{
-					_localValueSetter = BindingPropertyHelper.GetValueSetter(
-						_dataContextType,
-						PropertyName,
-						convert: true,
-						precedence: DependencyPropertyValuePrecedences.Local
-					);
-				}
-			}
-
-			private void BuildAnimationFillingValueSetter()
-			{
-				if (_animationFillingValueSetter == null && _dataContextType != null)
-				{
-					_animationFillingValueSetter = BindingPropertyHelper.GetValueSetter(
-						_dataContextType,
-						PropertyName,
-						convert: true,
-						precedence: DependencyPropertyValuePrecedences.FillingAnimations
-					);
-				}
-			}
-
-			private void SetSourceValue(ValueSetterHandler setter, object? value)
-			{
-				// Capture the datacontext before the call to avoid a race condition with the GC.
-				var dataContext = DataContext;
-
-				if (dataContext != null)
-				{
-					try
-					{
-						setter(dataContext, value);
-					}
-					catch (Exception exception)
-					{
-						if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Error))
-						{
-							this.Log().Error($"Failed to set the source value for [{PropertyName}]", exception);
-						}
-					}
-				}
-				else
-				{
-					if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Debug))
-					{
-						this.Log().DebugFormat("Setting [{0}] failed because the DataContext is null for. It may have already been collected, or explicitly set to null.", PropertyName);
-					}
-				}
-			}
-
-			private void BuildValueGetter()
-			{
-				if (_valueGetter == null && _dataContextType != null)
-				{
-					_valueGetter = BindingPropertyHelper.GetValueGetter(_dataContextType, PropertyName, _precedence, _allowPrivateMembers);
-				}
-			}
-
-			private void BuildPrecedenceSpecificValueGetter()
-			{
-				if (_precedenceSpecificGetter == null && _dataContextType != null)
-				{
-					_precedenceSpecificGetter = BindingPropertyHelper.GetValueGetter(_dataContextType, PropertyName, _precedence, _allowPrivateMembers);
-				}
-			}
-
-			private void BuildSubstituteValueGetter()
-			{
-				if (_substituteValueGetter == null && _dataContextType != null)
-				{
-					_substituteValueGetter =
-						BindingPropertyHelper.GetSubstituteValueGetter(_dataContextType, PropertyName, _precedence ?? DependencyPropertyValuePrecedences.Local);
-				}
-			}
-
-			private object? GetSourceValue()
-			{
-				BuildValueGetter();
-
-				return GetSourceValue(_valueGetter!);
-			}
-
-			private object? GetSourceValue(ValueGetterHandler getter)
-			{
-				// Capture the datacontext before the call to avoid a race condition with the GC.
-				var dataContext = DataContext;
-
-				if (dataContext != null)
-				{
-					try
-					{
-						return getter(dataContext);
-					}
-					catch (Exception exception)
-					{
-						if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Error))
-						{
-							this.Log().Error($"Failed to get the source value for [{PropertyName}]", exception);
-						}
-
-						return DependencyProperty.UnsetValue;
-					}
-				}
-				else
-				{
-					if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Debug))
-					{
-						this.Log().DebugFormat("Unable to get the source value for [{0}]", PropertyName);
-					}
-
-					return null;
-				}
-			}
-
-			private void BuildValueUnsetter()
-			{
-				if (_valueUnsetter == null && _dataContextType != null)
-				{
-					_valueUnsetter = _precedence == null ?
-						BindingPropertyHelper.GetValueUnsetter(_dataContextType, PropertyName) :
-						BindingPropertyHelper.GetValueUnsetter(_dataContextType, PropertyName, precedence: _precedence.Value);
-				}
-			}
-
-			internal void ClearValue()
-			{
-				BuildValueUnsetter();
-
-				// Capture the datacontext before the call to avoid a race condition with the GC.
-				var dataContext = DataContext;
-
-				if (dataContext != null)
-				{
-					_valueUnsetter!(dataContext);
-				}
-				else
-				{
-					if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Debug))
-					{
-						this.Log().DebugFormat("Unsetting [{0}] failed because the DataContext is null for. It may have already been collected, or explicitly set to null.", PropertyName);
-					}
-				}
-			}
-
-			private void BuildAnimationFillingValueUnsetter()
-			{
-				if (_animationFillingValueUnsetter == null && _dataContextType != null)
-				{
-					_animationFillingValueUnsetter = BindingPropertyHelper.GetValueUnsetter(
-						_dataContextType,
-						PropertyName,
-						precedence: DependencyPropertyValuePrecedences.FillingAnimations
-					);
-				}
-			}
-
-			internal void ClearAnimationFillingValue()
-			{
-				BuildAnimationFillingValueUnsetter();
-
-				// Capture the datacontext before the call to avoid a race condition with the GC.
-				var dataContext = DataContext;
-				if (dataContext != null)
-				{
-					_animationFillingValueUnsetter!(dataContext);
-				}
-			}
-
-			private void RaiseValueChanged(object? newValue)
-			{
-				ValueChangedListener?.OnValueChanged(newValue);
-			}
-
-			/// <summary>
-			/// Subscribes to property notifications for the current binding
-			/// </summary>
-			/// <param name="action">The action to execute when new values are raised</param>
-			/// <returns>A disposable to be called when the subscription is disposed.</returns>
-			private IDisposable SubscribeToPropertyChanged()
-			{
-				var disposables = new CompositeDisposable(_propertyChangedHandlers.Count + 1);
-				if (SubscribeToNotifyCollectionChanged(this) is { } notifyCollectionChangedDisposable)
-				{
-					disposables.Add(notifyCollectionChangedDisposable);
-				}
-
-				for (var i = 0; i < _propertyChangedHandlers.Count; i++)
-				{
-					var handler = _propertyChangedHandlers[i];
-
-					var valueHandler = new PropertyChangedValueHandler(this);
-
-					var handlerDisposable = handler.Register(_dataContextWeakStorage!, PropertyName, valueHandler);
-
-					if (handlerDisposable != null)
-					{
-						valueHandler.PreviousValue = GetSourceValue();
-
-						// We need to keep the reference to the updatePropertyHandler
-						// in this disposable. The reference is attached to the source's
-						// object lifetime, to the target (bound) object.
-						//
-						// All registrations made by _propertyChangedHandlers are
-						// weak with regards to the delegates that are provided.
-						disposables.Add(() =>
-						{
-							var previousValue = valueHandler.PreviousValue;
-
-							valueHandler = null;
-							handlerDisposable.Dispose();
-							OnPropertyChanged(previousValue, DependencyProperty.UnsetValue, shouldRaiseValueChanged: false);
-						});
-					}
-				}
-
-				return disposables;
-			}
-
-			public void Dispose()
-			{
-				_disposed = true;
-				_propertyChanged.Dispose();
-			}
-
-			/// <summary>
-			/// Property changed value handler, used to avoid creating a delegate for processing
-			/// </summary>
-			/// <remarks>
-			/// This class is primarily used to avoid the costs associated with creating, storing and invoking delegates,
-			/// particularly on WebAssembly as of .NET 6 where invoking a delegate requires a context switch from AOT
-			/// to the interpreter.
-			/// </remarks>
-			private class PropertyChangedValueHandler : IPropertyChangedValueHandler, IWeakReferenceProvider
-			{
-				private readonly BindingItem _owner;
-				private readonly ManagedWeakReference _self;
-
-				public PropertyChangedValueHandler(BindingItem owner)
-				{
-					_owner = owner;
-					_self = WeakReferencePool.RentSelfWeakReference(this);
-				}
-
-				public object? PreviousValue { get; set; }
-
-				public ManagedWeakReference WeakReference
-					=> _self;
-
-				public void NewValue()
-				{
-					var newValue = _owner.GetSourceValue();
-
-					_owner.OnPropertyChanged(PreviousValue, newValue, shouldRaiseValueChanged: true);
-
-					PreviousValue = newValue;
-				}
-
-				public void NewValue(DependencyObject dependencyObject, DependencyPropertyChangedEventArgs args)
-					=> NewValue();
-			}
-		}
 	}
 }
 #endif
